Create plex_playlist_generator
Script to generate a playlist on a plex server that randomizes the series, but plays the episodes within that series in order. Allows the admin to select whether to randomize their own playlist, or the plex home users. Plays the episodes within the series in proper order for each selected user. Allows the creation of randomized movie playlist, TV Show playlist, or a mix of both TV and movie playlist. Each user will receive their own randomly generated shows and/or movies. Supports the selection of specific Library selections by name.